HYD-4726 Add as-value directive
There are directives that expect a value as input. When using a
directive like this inside an ng-repeat, it needs a value cooresponding
to each row displayed.

Given we have a stream, we can create a property. This property can be
combined with an optional through stream that will transform the final
value. Once the value is optionally transformed, it will be emitted on
the transcluded scope. This will allow direct assignment into a
directive.

  - Add as-value directive.
  - Add as-value tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>