HYD-4729 GUI: as-prop directive. update through to use parent context
In order to properly filter a stream being forked inside a ng-repeat,
we need to update through to take args. We'll rename it to transform.
transform should take the stream and an optional array of args.
This will allow it to function the same as the through binding does today.

  - Changed through to transform
  - changed attr type from =? to &?
  - added args =? binding
  - Updated tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>