HYD-4751 Update GUI such that ui-modules point to the correct location
Since we've moved our ui-module into a private repo the modules are now
being ziplocked and pulled down into a node_modules folder under
ui-modules. Due to this new structure the gulp-src-globs must be updated
to point to the correct location. Additionally, the conf paramters need
to be updated in the index.html such that they use conf.get('<varname>')
instead of conf.<varname>.

This patch will update the gulp-src-globs.json and index.html file in
the GUI private repo. Once this branch is merged in we can publish a new
version and the GUI can then be ziplocked in the chroma_ui_new
directory.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>