HYD-4752 Update gulp dependencies to be in the root GUI directory
The gulp dependencies are in the gulp folder but should be moved up into
the root directory (GUI). When gulp is run, it will go up to the parent
to look for the dependencies and when it finds them it will run the
specified gulp task.

This patch will update the package.json file in the GUI module's gulp
directory by removing all dev dependencies and bringing them to the
GUI's package.json.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>