HYD-4516 Implement Generic Lexer / Parser in JavaScript
We have a number of areas in the GUI where we could benefit from a generic Lexer / Parser.
By providing this, we abstract repetitive code into a reusable module
and make future areas needing a parser faster to implement.

Implement a generic Parser / Lexer that can handle a grammar, a list of rules and an output format and convert input to output.

  - Created parslely module.
  - Used combinatorial parsers inspired by parsec
    to build up a parser.
  - Wrote tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>