HYD-4760 Resolve issue where highland is not included in the GUI distribution
After publishing the GUI module we discovered that the highland dist
folder is missing from the packaged source. This dist folder and its
associated source should be included in the package source.

This patch will update the GUI module such that we no longer pull
highland through bower components and instead pull it through npm. Both
gulp-src-globs and karma.conf will be updated to reflect the change and
the existing highland directory in the bower_components directory will
be removed.

- Updated .gitignore to ignore the highland package.json
- deleted the highland package.json

Signed-off-by: Will Johnson <william.c.johnson@intel.com>