HYD-4768 Update socket worker such that the timeout receives an error arg when necessary.
When testing the GUI we ran into an issue where an error was not being
passed to an event handler. Since we are using $timeout we need to make
sure that the error is passed as the fourth arg.

This patch will update the socket worker and its associated test to
ensure that the error object is passed into the handler when the error
occurs.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>