HYD-4229 Help not available when not logged in
Help is only available in the top menu bar when you are logged in.
This seems incorrect to me as a logged out user may need to reference help,
and I can't think of any way Help would contain sensitive information.
Would like to open discussion on if this should be unrestricted.

  - Removed restriction on fs-admin for help in GUI.

Signed-off-by: Joe Grund <joe.grund@intel.com>