HYD-4525 Update status screen
Problem
The status screen is currently of limited use.
Users can currently select from All, Alerts, Commands, and Events and
page between those types.
They can also view dismissed items and dismiss all items.
Anything beyond that is not achievable even though we have a queryable
/alert endpoint that is quickly becoming more searchable.

Solution

Update the status screen to be queryable a-la a Jira issue page. By
doing this, we limit viewable information to what the user can think up
instead of predefined views.
We can also pass data into the screen as a querystring. By doing so, we
can pre-filter the screen to display the data we want before we get
there.

Dismiss / Dismiss all will disappear as the are not needed when you can
select the records you are interested in.

  - Update status screen to be queryable.
  - Split out nested components into directives.
  - Create sub-view hiearchy so table can
    be refreshed by a query string route change.
  - Add tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>