HYD-4802 GUI@3.0.16: create multiStream
In order to support actionable alerts, we need a mechanism that combines
multiple streams and returns their values as an array.

t's important to note that the multiStream itself returns a stream so
further processing can be performed.

Also, multiStream acts like Promise.all in that no values will be
emitted until all streams pass through data.
Errors will be propagated the same as any other streams as
well.

  - Create multiStream
  - Add tests

Signed-off-by: Joe Grund <joe.grund@intel.com>