HYD-4663 Create chart legend component
We want our charts to be broken up into re-usable components that will
allow us to put them together as quickly as possible. Charts are made up
of several components including the grid, the lines / bars representing
the data, axis, ticks on axis, legend, etc. In this ticket, we want to
separate out the legend such that it can be re-used throughout multiple
charts.

This patch will add a get-legend.js and an associated test that can be
used to add a legend to a chart.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>