HYD-4830 Create as-stream directive
Now that action-dropdown is going to use a stream, we need a val to
convert singular values into a stream at runtime.

Create as-stream to achieve this. as-stream is the dual of as-value.
It will be given a $scope and will watch the val property on that scope.
It will take any changed value and push it into a writable stream it
creates.
It will expose the stream as str on a transcluded scope.

  - Create as-stream directive.
  - Create DOM tests to cover common usage.

Signed-off-by: Joe Grund <joe.grund@intel.com>