HYD-4870 Create route stream
We have many pages that are made up of nested segments using
angular-route-segment. While angular-route-segment provides a way reload
a route on change, it does not provide a way to re-push resolved
functions into a route on route change, without a re-render.

Because of this, we should create a route-stream. It's purpose is to
push any route changes onto a stream if they are not redirects. This
will allow us to listen on route changes and react accordingly in our
controller code.

  - Add route-stream module.
  - Add tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>