HYD-4513 Utilize Natural Sort Module to sort OSTs
The list of OST's is being displayed in a lexical order as opposed to a
natural order (see ATP-15).

By utilizing a natural-sort we can make sure the OSTs are in order
regardless of how the are added to the system. To do this, we need to
update the get-read-write-heatmap-stream and
get-ost-balance-transformer-stream such that the data is sorted using
localeCompare. This patch will update both get-ost-balance-stream and
get-read-write-heat-map-stream.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>