HYD-4888 Create chainL1 parser
In order to parse in statements, which our old version of tastypie only
accepts like:

a_in=1&a__in=2&b__in=3

We should reduce consecutive in statements to an array. We can do this
using a chainL1 parser.

It will take a parser and operation HOF.

parser output will be used to feed the operation, which will have an arity of
2.

This will end up being a left associative parser, so it could be used
to model things like addition as an example.

  - Add chainL1 service.
  - Add tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>