HYD-4914 GUI: Dehydrate lnet_config host relation
The lnet_configuration endpoint recently changed to do a full relation
of host. We need just the resource_uri of a host to link it with LNet
status on the server screen.

Use the new dehydrate__=false functionality to have the API reduce the
host to a resource_uri.

  - dehydrate host.
  - update test.

Signed-off-by: Joe Grund <joe.grund@intel.com>

Conflicts:
	source/chroma_ui/iml/server/server-streams.js
	test/spec/iml/server/server-streams-test.js