HYD-4943 Update qs-to-input-parser to reduce *__in query strings
Tastypie only accepts array like syntax as:

  severity__in=ERROR&severity__in=WARNING
This needs to be parsed to:

  severity in [ERROR, WARNING]

In order to do this, we need to reduce consecutive usages of __in.
We need to make sure to group by lhs value and only reduce the
closest value to the left so it maintains the same parse structure
as the input.

  - Update qs-to-input-parser to handle reduction of __in statements.
  - Update tests to add more cases.

Signed-off-by: Joe Grund <joe.grund@intel.com>