HYD-4952 Create deferred action dropdown directive
We can have many action dropdowns on the status screen depending on the
number of issues.

If we were to load the intersection of all affected resources across
visible rows at once, it could be a lot for the API / GUI to handle.
Instead, defer loading the affected list until the user mouses over the
action dropdown.

Then, we can display a loading spinner while the resources are being
fetched in parallel.

Finally, we can re-enable the button when all resources have loaded.

  - Created deferred-action-dropdown directive.
  - Added tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>