HYD-4962 Status button should show active alerts when yellow or red
As a user, I would expect the status button to show me the issues with
my system when I click it. In order to do so, we should dynamically
build the link depending on our health.

If we are GOOD:

/ui/status/

If we are not GOOD:

/ui/status/?severity__in=WARNING&severity__in=ERROR&active=true

This will ensure the user will see what they expect when clicking.

  - Add dynamic link building.
  - Remove old code.

Signed-off-by: Joe Grund <joe.grund@intel.com>