HYD-4972 Update jscs and jshint in GUI to account for spaces between function and parenthesis
Our style guides indicate that a space should be included between a
function and the parentheses as well as the brace. For example:

// Good:
function funcName ()
{ ... }
function () { ... }
// Bad:
function funcName()
{ ... }
function() { ... }
We need to update the jshint and jscs to check for this and fix any
existing style issues.

This patch will update the .jscsrc file so that jscs will make sure
function declarations and function expressions both contain a space
between the open paren as well as the brace.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>