HYD-4168 Update angular-route-segment to handle middleware
Currently the app is monkey patching the router to add in a
"segmentAuthenticated" section such that the users credentials are
checked when navigating to a route. This is not the best way to do this
and we've identified that the check is run in parallel along with the
other local resolves for the route. Instead, we need a solution where
each route can specify a list of middleware that will be run when
navigating to each route. Unfortunately, angular-route-segment does not
provide this capability, which is why we forked the repo into @intel-js
and added the functionality in 1.6.1.

With the ability to add middleware to the angular-route-segment, this
patch will add a middleware processor that will receive the list of
middleware for a route and use the injector to invoke each middleware.
Each middleware will return a promise and if all promises resolve the
route will proceed and THEN the local resolves will be processed. If any
of the middleware items reject the route is aborted without processing
any local resolves and if a path is specified as a parameter to the
rejected promise the user will navigate back to the specified path.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>