HYD-5006 Create page-visibility module
Create a module that utilizes the page visibility API to create and
destroy an input stream dynamically as the page goes from foreground to
background.

This will be done by returning an empty passthrough stream that gets
populated by data from the input stream when the page is visible.
When the page is hidden, a predefined empty value is pushed through the
stream to flush it.

  - Add module to start/stop a stream when visibility changes.
  - Add tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>