HYD-5193 GUI: Add chartcompiler directive
We need a way to compile a chart by using an HTML attribute. Previously
the logic to compile a chart was tied into the dashboard charts. Now we
can extract it to something like:

<div chart-compiler chart="::chart"></div>

by doing so, the chart represented by chart gets appended to the element
containing the attribute.

This also allows us to clean up some existing code that was doing the
chart compilation process inside a directive link function. In other
words, the process becomes declaritive.

  - Add chart-compiler directive
  - Refactor charts-container to use chart-compiler directive.
  - Update to ES6.

Signed-off-by: Joe Grund <joe.grund@intel.com>