HYD-5243 GUI: Chart should be immutable
The compile function of a chart should be largely immutable. By this I
mean we should be able to call the compile function and receive a new
chart DOM fragment to place into the DOM tree. We should be able to call
compile again, either before or after destroying the existing chart
fragment and receive another usable fragment.

Currently, the DOM node is not being cloned so we are actually reusing
the same fragment on the second call. In addition, we do not fork or
create a property off the stream so the second time we create a fragment
it will error. We need to:

1. Clone the chart fragment in the compiler function.
2. add a property to the incoming stream to chartCompiler.
3. Add a destroy method to the compiler so we can destroy the source
stream when the chart is truly done with it.

  - Make chart clone DOM node.
  - Make chart create stream property.
  - Add tests.
  - Update code to destroy chart.

Signed-off-by: Joe Grund <joe.grund@intel.com>