HYD-5227 Server actions should be disabled with description when not available
The "Install Updates" button is enabled on the configure/server page
when it should not be.
There are four states at play for this button. They are:

1) All servers are member of an active fs.
2) No servers have updates available.
3) Some combination of 1 and 2 that amounts to all servers.
4) None of the above

In cases 1-3 the button should be disabled. In case 4 it should be
enabled.

In addition to this, the buttons may be enabled but some of the row
level toggles might be disabled. In this case we need to know what type
of action we are performing and display why that particular toggle is
disabled.

  - Update server-actions-factory and server.html to
    account for new states.
  - Refactor some server screen code + tests to ES6.
  - Fix bug where command modal did not open in
    confirm-server-action-modal-ctrl.

Signed-off-by: Joe Grund <joe.grund@intel.com>