HYD-5284 Update page-visibility to take a timeout
The page-visibility service currently takes functions that are called
when the page is hidden and shown respectively. In addition to that, we
can add a timeout which will only call onHide after n ms have elapsed.
By doing this, we can avoid re-rendering things when someone flips back
and forth between tabs.

  - Move page-visibility to ES6 + tests.
  - Add timeout param.

Signed-off-by: Joe Grund <joe.grund@intel.com>