HYD-5282 GUI: babel should not compile non-ES6 code
Babel is currently compiling much more then it needs to in our gulp
tasks. It should only be compiling ES6 code. We use an ignore blacklist
to decide what should not be compiled. Instead of that, switch to an
only whitelist as it's much smaller and unlikely to change as much.

  - Change from blacklist to whitelist.
  - Add singlequote output for ng-annotate.

Signed-off-by: Joe Grund <joe.grund@intel.com>