HYD-5337 GUI: remove usages of pathSegList
Chrome 48 was released yesterday. As part of it's release it removes the
pathSegList attribute which gets nicely formatted path information. This
is causing our tests to fail as we rely on this attribute for coordinate
information.

Replace pathSegList usages with getAttribute('d') followed by some
parsing to gain the same structure.

  - Change usages of pathSegList to custom parsing.
  - Update intel-math to 3.0.0.
  - Update usages of intel-math to rely on ES6 imports.

Signed-off-by: Joe Grund <joe.grund@intel.com>