HYD-5438 GUI: Implement System.js module loading
Background
Modules are supported in the ES6 (ES2015) spec. This provides developers with first-class means to create reusable code. Previously, many competing specs and strategies (AMD, UMD, CommonJS, Globals, Angular modules) have been devised as a user-land replacement for this hole in the language.
This has been deficient for a number of reasons:

1. The choice of module may discourage code reuse in different environments. (Ex: angular modules are not portable to a non-angular codebase)
2. Certain strategies are not reliable for resolving dependencies (Ex: UMD defaults to globals in a browser context.

The GUI repo was created before ES6 / Babel was an option. As such it currently uses Angular modules. This makes our code largely non-portable and also makes it difficult to consume modules with dependencies in the GUI.

We have added babel transpilation into our build step. We currently compile to UMD modules, which has the aforementioned global limitations.

This ticket is to port module loading and bundling to the System.js format instead. This has the following advantages.

1. The dependency graph is statically analyzed. This means dependencies can be determined during transpile time, a specific ordering of script tags in index.html is no longer needed.
2. Modules can be broken into related bundles and lazy loaded on demand. This means different pages can be split into bundles; the whole app does not need to be delivered up front.
3. Because the dependency graph is statically analyzed, modules with dependencies will resolve correctly. This also means no more globals. Another big win of this is modularity; we can start to split GUI up into small pieces that are easily reusable in different environments.
4. Also because of static analysis, we can use tools that do tree-shaking to eliminate dead code at the module level. This will reduce the payload of the app delivered over the wire.
5. Development is simplified as System.js loads deps via AJAX during dev. This means we do not need to re-write / reload index.html (which is cached by view-server) every time we make a change.
6. the System global object was a standard in ES2015. It was removed from the spec, but seems like it may appear in a later version. If that happens, we will have a polyfill in place that should just work with native implementation.
7. Future tools (Angular 2) all work best with an ES6 module compatible environment.
8. IDEs can use the import information to provide information about imported modules used in code.

Implementation

All angular dependencies, currently resolved by angular, now need to be ES6 import statements that can be statically analyzed. This involves adding explicit imports to each file needed to stitch the dependency graph together.

Because this approach uses static analysis. The final implementation should be stable. I.E. we should know early on if an import is missing / not pathed correctly.

  - Update all source files to be ES6 modules.
  - Stitch all modules together.
  - Remove internal big-differ module.

Signed-off-by: Joe Grund <joe.grund@intel.com>