HYD-4807 Saving corosync configuration does not provide a response
PUTs to /corosync_configuration return a 204 with a null body.
This is because there is a special rule that any PUT with
a state property is for a state change.
In other words this property can not be present if
you are trying to modify some other attribute.

Update the PUT to /corosync_configuration exclude state,
and only include network_interfaces, mcast_port and id.

  - Update server-detail-resolves to include network_interfaces property.
  - Update PUT to only send the needed data.

Signed-off-by: Joe Grund <joe.grund@intel.com>