HYD-5593 GUI: Can no longer see HSM coordinator status until after a copytool is created
You used to be able to check on the status of the HSM coordinator on the
HSM page prior to adding a copytool. On the new page, you now cannot see
it until after adding a copytool.

This is because we check for copytools before showing the cdt-status
component. There appears to be no good reason for this change as this
component works fine without any copytools (at least with the
simulator).

  - Remove blocking cdt-status when there are no copytools.

Signed-off-by: Joe Grund <joe.grund@intel.com>