HYD-5440 GUI: precompile app on npm publish
We currently build the GUI for production in Jenkins. The process looks
something like this for JS files:

source -> transpile (babel) -> concat -> uglify -> generate sourceMaps

Doing this on the CI server was an artifact of not wanting to check
compiled code into source control. Now that we have split out into a
separate Github repo that is backed by modules on Artifactory, we can
build on prepublish for new module releases. We currently do this for
our other ES6 -> ES5 modules, so it makes sense to do the same for GUI.

This has the following advantages:

1. No need for CI to spend time doing building.
2. We lock the bits used to build a version when we publish a module.
3. We can move some dependencies to devDependencies, where they should be.

Implementation

1. Add a prepublish task that does a gulp build.
2. Add dest dir to .gitignore.
3. Add a .npmignore that does not have dest in it.
4. Update Jenkins build to not compile GUI.

Signed-off-by: Joe Grund <joe.grund@intel.com>