HYD-5687 GUI: implement qs-to-old-qs-parser
Currently we parse querystrings containing list clauses like:

x in [1,2,3] -> x_in=1&xin=2&x_in=3.

This is actually difficult to implement when having to consider other
alternatives.

Newer versions of Tastypie support a more compact list syntax:

x__in=1,2,3

This syntax is actually much easier to parse to. The plan of attack is:

1) parse to and from new list syntaxes even though they are not
supported yet in the API.

2) Add a qs-to-old-qs-parser That will convert the new style lists to
the old style list syntax. This parser will convert before we send data
to the API. It can remain in place until the API is updated to a version
that supports the new format.

  - Implement qs-to-old-qs-parser + tests.
  - Add intel-parsely so it's loaded.
  - Add babel-plugin-transform-object-rest-spread.
  - Update querystrings being passed to API to use new syntax parsed
    to old syntax.