HYD-5642 GUI: update common status searches to use new parsers
With HYD-5641 and HYD-5639 we can correct and shorten the query strings i
used in the common-status-searches component.

This is a simple change that just involves updating a few query strings.

  - Update HTML + tests.

HYD-5641 GUI: Implement parser for ends with expression

The status parser needs a way to parse matches on the end of a string.

This can be done with a endswith expression:

The syntax will look like:

FIELD ends with VALUE

which will serialize to:

field__ends_with=VALUE

  - Implement ends with parser + tests.

HYD-5639 GUI: Implement parser for contains expression

The status parser needs a way to parse substring matches.

This can be done with a contains expression:

The syntax will look like:

FIELD contains VALUE

which will serialize to:

field__contains=VALUE

  - Implement contains parser + tests.

  - In addition, both parsers have been simplifed by using a newer style list syntax.

Signed-off-by: Joe Grund <joe.grund@intel.com>