HYD-5440 GUI: precompile app on npm publish
Background
We currently build the GUI for production in Jenkins. The process looks
something like this for JS files:

source -> transpile (babel) -> concat -> uglify -> generate sourceMaps

Doing this on the CI server was an artifact of not wanting to check
compiled code into source control. Now that we have split out into a
separate Github repo that is backed by modules on Artifactory, we can
build on prepublish for new module releases. We currently do this for
our other ES6 -> ES5 modules, so it makes sense to do the same for GUI.

This has the following advantages:

  * No need for CI to spend time doing building.
  * We lock the bits used to build a version when we publish a module.
  * We can move some dependencies to devDependencies, where they should be.

Implementation

  * Add a prepublish task that does a gulp build.
  * Add dest dir to .gitignore.
  * Add a .npmignore that does not have dest in it.
  * Update Jenkins build to not compile GUI.

  - Create a dist dir for built code.
  - Do not run prepublish in test-gui.sh
    npm i will do this.
  - Move all deps to devDependencies
    The app is built on prepublish.
  - Add ignored dirs to dotfiles.

Signed-off-by: Joe Grund <joe.grund@intel.com>