HYD-5708 GUI: use preversion to build for production
We are currently using prepublish to build GUI for deployment in 3.1.

This would work nicely, except that npm runs prepublish as well on
dev-style installs.

Switch to using preversion to build GUI for deployment. By doing so, we
can do a developer style install without having to rebuild the app.

As part of this, we should also create a zsh script that will do a
version bump and publish for us.

  - Switch from prepublish to preversion.
  - Make sure we build during test.

Signed-off-by: Joe Grund <joe.grund@intel.com>