HYD-5713 GUI: create store
Background

The GUI should have a single store that keeps the state of the app.

By doing this, we decouple state from components and allow components to
be rendered at their last known state upon reuse.

Additionally, we can decouple datasources from components. The store
will serve as a proxy that will both be seeded with data on page load
and will keep the last known state as pushed from the realtime service.

This will largely be an implementation of the redux pattern, but using
reactive programming.

Implementation

Create a store on page load. This store will be seeded by data coming
from CACHE_INITIAL_DATA. Further updates will be dispatched by a
WebSocket as changes occur for relevant endpoints.

The store should be immutable. Any changes should create a new reference
to the data, never mutate it.

Only one store should be created; it is the central point for all state
in the app.

In order for subscribers to the store to get the most recent state on
subscription, we will use asProperty.

We will use scan instead of a plain reduce function so state
can be emitted as it occurs.

  - Add store module.
  - Add tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>