HYD-5730 GUI: store should emit without backpressure
The store needs to be updated to emit values to multiple subscribers
without backpressure.

Currently we are using a forking strategy (through add-property), where
new subscribers cause a fork that get's latest data.
This can cause a deadlock in that no data will flow unless both streams
pull. The issue occurs when one stream depends on another stream pulling
first.

Instead of this, we can do a simple broadcast strategy. This means
viewers can subscribe to changes on the main broadcaster stream. We do
not apply backpressure here, so all streams will get the latest data
when they subscribe. Additionally, we do not need to buffer changes in
the broadcast stream.

  - Change strategy to use broadcast instead of properties + forks.
  - Add tests to cover changes.
  - Add typing to socket-stream.

Signed-off-by: Joe Grund <joe.grund@intel.com>