HYD-5397 Make profile selection more user friendly
When installing on EL7.2 the first profile the user should select is the
second in the list, it would be more friendly if the GUI switched to
this profile once it goes green.

This patch will solve the above problem by sorting the profiles such
that invalid profiles are pushed to the bottom of the list. This will
maintain order but will ensure that only selectable items remain at the
top of the list.

- select-server-profile-step.js - Sorted the profiles such that invalid
  profiles appear at the bottom of the list.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>