HYD-5740 GUI: implement extract-api-filter
It is useful to convert a resource_uri to an id in an angular template.
In other words a transform like:

'/api/host/1' -> '1'

Once we have this, we can route to a resource detail page easily.

We can implement this as a filter, but filters are expensive due to the
amount of times they are called during digest.

We can overcome this by using fp.memoize. By doing this we refer to the
cached value and do no use any cpu time recalculating the same value.

  - Add extract-api-filter + test.

Signed-off-by: Joe Grund <joe.grund@intel.com>