HYD-5746 GUI: create target reducer
We can cache the data for the /target endpoint by implementing a reducer
that is used by store.

By doing this, we gain instant loading of pages using /target.

Initially we will cache the entire endpoint as it's transfered on page
load from view-server. This should scale to thousands of entries, maybe
more if we prune some unneeded data with json-mask. Longer term, we will
want to fetch a reasonable subset if scale becomes an issue.

  - Add target-reducer + tests.
  - Add action type to store.

Signed-off-by: Joe Grund <joe.grund@intel.com>