HYD-5720 GUI: update action dropdown to disable on no data
The /storage_resource endpoint contains records that do not have the
available_actions field that other resources do. As a side affect, when
rolling the mouse over an actions button an exception will be thrown.
Rather than throw an exception, the app should be more resilient and
should change the button to say "No Actions" instead of displaying an
error.

This patch will filter out all records that do not have locks or do not
have the actionsProperty in the record. After filtering, only records
containing the actionsProperty and are not locked will remain. The
action-dropdown template will then be updated to display a "No Actions"
button in the case where there are no records after the filter. If any
records exist, the standard actions button will be displayed.

- action-dropdown.js - Filter data on record locks and the actions
  property
     - Move tooltipPlacement and actionsProperty into the controller to
       avoid timing issues
- action-dropdown.html - Add a "No Actions" button that displays if
  there are no records
    - Update the Actions button to display only if there are records or
      if data has not yet been received.
- action-dropdown-test - Update tests

Signed-off-by: Will Johnson <william.c.johnson@intel.com>