HYD-5759 GUI: add reducers and streams for alerts and jobs
Many records in the GUI are associated with alerts and jobs. Currently,
we create a stream on demand for each of these, and tear down when
navigating away.

Instead of this, we can create the streams once on page load and use the
store to keep their state. By doing this, the loads become synchronous.
We should also pre-fetch this data in the view-server. That can be a
separate ticket.

  - Add reducers, streams + tests for alerts and jobs.

Signed-off-by: Joe Grund <joe.grund@intel.com>