HYD-5775 Update host call to use the store in server-detail-resolves
Now that the call to /hosts is being stored in the store, we should no
longer establish a socket connection to /hosts in
server-detail-resolves. Instead, host data should be retrieved from the
store.

This patch will update server-detail-resolves such that a socket
connection is no longer used to retrieve host data. Instead, data will
be pulled from the store in real time.

- server-detail.html - Removed comments
- server-detail-resolves.js - Update the socket call to /hosts to
  retrieve data from the store and filter on the route id
- server-detail-resolves-test.js - Update tests to ensure we pull server
  data from the store, remove socket call test, and ensure that the data
  is filtered correctly

Signed-off-by: Will Johnson <william.c.johnson@intel.com>