HYD-5800 GUI: Use more space if available for confirmation dialog text
Sometimes actions have many confirmations steps. Currently we place them
in a well that has a max-height of 200px and overflow: scroll. Because
of this, steps that overflow the well are below the scroll area. This
problem is much worse on OS X where scroll bars are hidden by default.

Remove the max-height: 200 and overflow: scroll rules so all steps will
be displayed. In the case the modal is larger than the screen height the
entire window will scroll, which is not an issue.

  - Remove rules that limit height and overflow on well.

Signed-off-by: Joe Grund <joe.grund@intel.com>