HYD-5789 Link filesystem to filesystem detail page for admins on dashboard page
When drilling into a filesystem's target on the dashboard page, the
target's name is displayed as regular text when the filesystem is active
and shows as — when the filesystem is stopped. It would be more user
friendly to link the target name to the server detail page associated
with that target when the filesystem is active. Regular text should be
displayed instead of the link when the filesystem is stopped and — is
displayed.

This patch will update the dashboard page such that the target name is
linked when the name is available and the text "---" is shown without a
link when the name is not provided.

- Updated target-dashboard.html such that the link is displayed when the
  active host is not null and the text "---" is displayed when active
  host is null
- Added the store to target-dashboard-module.js
- Updated targetDashboardTargetStreamFactory to pull the target stream
  from the store and filter on the targetId
- Fixed lint issue in lnet-configuration-reducer.js
- Updated target-dashboard-resolves-test.js to verify that target data
  is pulled from the store

Signed-off-by: Will Johnson <william.c.johnson@intel.com>