HYD-5806 GUI: Link filesystem name on dashboard target row
When drilling down to a target on the dashboard page the filesystem name
is displayed as text. It would be more user friendly if the filesystem
name was linked to its corresponding filesystem detail page.

This patch will update the target row such that the filesystem name will
be linked to its corresponding filesystem detail page if the user has
admin privileges. If the user does not have admin privileges, the
filesystem name will be rendered as text only.

- Updated target-dashboard.html to render the filesystem as a link if
  you are logged in as admin and as text if you are not logged in as admin.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>