HYD-5834 GUI: Create to-date filter
We need a way to convert strings to date objects. Using a date filter to
do this would allow us to cast a string to a date within a template.

This patch will add a to-date filter that can be used to cast
strings to Date objects within a template file.

- Created to-date-filter-module.js
- Created to-date-filter.js filter
- Added to-date-filter-test.js test

Signed-off-by: Will Johnson <william.c.johnson@intel.com>