HYD-5858 Store: select streams receive updates for any dispatch
There is a bug in the store where any dispatch to the store would cause
all select streams to get a new token; even though their view of the
store was unchanged.

Add a filter to select that removes tokens if they are the same
reference as the last thing that passed through the stream. We can use
reference checks because any dispatch to the store is guaranteed to
result in a new reference.

  - Bump fp to 4.2.3 which contains fix
    for using tap as a functor.
  - Update select to only send changed tokens
    when compared to last send.
  - Add test to catch regressions.

Signed-off-by: Joe Grund <joe.grund@intel.com>