HYD-5868 GUI: Job stats page allows a minimum window of 16 seconds, data slices are 10 seconds
The ui/dashboard/jobstats page shows no data for any window size of 15
seconds or less. Since data slices can be 10 seconds it may show no data
when there is data present. Data windows of at least 10 seconds, if not
less, should be supported.

This issue was fixed in GUI/3.0.0.0 (HYD-5832). The purpose of this
patch is to bring the same fix from HYD-5832 into GUI/master. This patch
will update the get-read-write-heat-map-chart.js file such that a
minimum window of 16 seconds is set for job stats. The corresponding
test file will be updated to verify this.

- get-read-write-heat-map-chart.js - updated to ensure minimum window of
  16 seconds for job stats
- get-read-write-heatmap-chart-test.js - added new test with a window of
  4 seconds and expect to get a result back with a 16 second window.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>