HYD-5926 GUI: Create resettable group component and resetter directive
It would be nice to be able to reset all fields in a form back to their
initial value when canceling a form submission. For example, if we have
a form with three text fields, that initially have a value of 1, 2, 3
respectively, and we update the fields to be 5, 6, 7 and click a reset /
cancel button, the text box values should go back to 1, 2, 3.

A resettable-group component and resetter directive will be added to
this patch . The resettable-group component requires the FormController,
allowing it to communicate with the form that it is part of. The
resetter directive can be placed on any button, link, etc, and will
notify the resettable-group component when the form should be reset.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>