HYD-5813 Create auto-complete component
Background

We would like to have the experience that the user types some data into
parsely-box and based on their cursor position, get some relevant
suggestions.

Architecture

There needs to be three communicating components to make this possible:

completionist

Responsible for coordinating events between child components. This will
have built in EventEmitter type methods. This component should not hold
state; it really just serves as a dispatcher.

completionist-model-hook

Responsible for interacting with the input field and dispatching events
upon certain occurrences.
Also responsible for updating the view when a new autocomplete value is
suggested. This component basically serves as the bridge between the
input field and the code.

completionist-dropdown

Responsible for the dropdown component that appears when the user types
something pertaining to a suggestion. Translates events from the input
field to actions taken on the dropdown. Dispatches any suggested items.

  - Implement completionist components + tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>